Feat [#64] 참여현황 버튼 구현
