+++ PINGLE-iOS/PINGLE-iOS/Presentation/Home/Components/ParticipantButton.swift
@@ -7,14 +7,120 @@
 
 import UIKit
 
-class ParticipantButton: UIButton {
+import SnapKit
+import Then
 
-    /*
-    // Only override draw() if you perform custom drawing.
-    // An empty implementation adversely affects performance during animation.
-    override func draw(_ rect: CGRect) {
-        // Drawing code
+final class ParticipantButton: UIButton {
+    
+    let participantsLabel = UILabel()
+    let rightArrowImageView = UIImageView()
+    let countStackView = UIStackView()
+    let currentParticipantsLabel = UILabel()
+    let slashLabel = UILabel()
+    let totalParticipantsLabel = UILabel()
+    let completeLabel = UILabel()
+    
+    // MARK: - Function
+    // MARK: LifeCycle
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setUI()
     }
-    */
-
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+}
+// MARK: - extension
+extension ParticipantButton {
+    // MARK: Layout Helpers
+    private func setUI() {
+        setStyle()
+        setLayout()
+    }
+    
+    private func setStyle() {
+        self.do {
+            $0.makeCornerRound(radius: 40.5)
+            $0.backgroundColor = .grayscaleG10
+        }
+        
+        participantsLabel.do {
+            $0.text = StringLiterals.Home.Detail.participantsTitle
+            $0.textColor = .white
+            $0.font = .captionCapMed12
+        }
+        
+        rightArrowImageView.do {
+            $0.image = ImageLiterals.Home.Detail.icParticipantArrowActivate
+        }
+        
+        completeLabel.do {
+            $0.text = StringLiterals.Home.Detail.complete
+            $0.textColor = .grayscaleG06
+            $0.font = .subtitleSubSemi16
+            $0.isHidden = true
+        }
+        
+        countStackView.do {
+            $0.axis = .horizontal
+            $0.spacing = 0
+        }
+        
+        currentParticipantsLabel.do {
+            $0.text = "1"
+            $0.font = .titleTitleSemi30
+        }
+        
+        slashLabel.do {
+            $0.text = StringLiterals.Home.Detail.slash
+            $0.textColor = .grayscaleG06
+            $0.font = .titleTitleSemi30
+        }
+        
+        totalParticipantsLabel.do {
+            $0.text = "10"
+            $0.textColor = .grayscaleG06
+            $0.font = .titleTitleSemi20
+        }
+    }
+    
+    private func setLayout() {
+        self.addSubviews(participantsLabel,
+                         rightArrowImageView,
+                         completeLabel,
+                         countStackView)
+        
+        countStackView.addArrangedSubviews(currentParticipantsLabel,
+                                           slashLabel,
+                                           totalParticipantsLabel)
+        
+        self.snp.makeConstraints {
+            $0.width.height.equalTo(81)
+        }
+        
+        participantsLabel.snp.makeConstraints {
+            $0.top.equalToSuperview().inset(15)
+            $0.leading.equalToSuperview().inset(15)
+        }
+        
+        rightArrowImageView.snp.makeConstraints {
+            $0.centerY.equalTo(participantsLabel)
+            $0.leading.equalTo(participantsLabel.snp.trailing).offset(1)
+        }
+        
+        completeLabel.snp.makeConstraints {
+            $0.top.equalTo(participantsLabel.snp.bottom).offset(8)
+            $0.centerX.equalToSuperview()
+        }
+        
+        countStackView.snp.makeConstraints {
+            $0.top.equalTo(rightArrowImageView.snp.bottom)
+            $0.bottom.equalToSuperview().inset(15)
+            $0.centerX.equalToSuperview()
+        }
+        
+    }
+    
 }
